[WARNING][Database Change] Change resource_size type from INTEGER to NUMERIC

+++ include/resto/RestoModel.php
@@ -130,7 +130,7 @@ abstract class RestoModel {
         ),
         'resourceSize' => array(
             'name' => 'resource_size',
-            'type' => 'INTEGER'
+            'type' => 'NUMERIC'
         ),
         'resourceChecksum' => array(
             'name' => 'resource_checksum',
